fix IMPORTS of schemas with STARTS
The ShapeExpr associated with a START= directive gets added to the
rules with a null hash key. Schemas which imported another schema
which had a START= lead to a hash conflict on null, causing all but
the last START ShapeExpr to have null IDs.

e.g. start2RefS1-IstartS2, start2RefS2-IstartS2

The rigth way to fix this may be to ensure that START= gets something
other than a null hash key, but you'd still have to prevent collisions.